worked in the _Layout view to make menu better

+++ MLandfill/tblInterestCharge.cs
@@ -14,7 +14,7 @@ namespace MLandfill
     
     public partial class tblInterestCharge
     {
-        public int IntAmount { get; set; }
+        public decimal IntAmount { get; set; }
         public int IntDaysOverdue { get; set; }
     }
 }
